+++ src/display.cpp
@@ -1798,9 +1798,7 @@ void display_show_image(uint8_t *image_buffer, int data_size, bool bWait, bool b
         bbep.freeBuffer();
     }
 #else
- {
     fastepd::update(bWait, bSkipClear, iUpdateCount, iTempProfile);
- }
 #endif
     iUpdateCount++;
     Log_info("display_show_image end");
