+++ src/displays/fastepd.h
@@ -1,12 +1,15 @@
 #pragma once
 
-// Unified FastEPD entry point for BOARD_X_CLASS targets.
-// Callers use fastepd::{init,update} and bbep; board selection stays here.
+#ifdef BOARD_X_CLASS
 
-#ifdef BOARD_TRMNL_X
-#include "fastepd_x.h"
-namespace fastepd = fastepd_x;
-#elif defined(BOARD_X_CLASS)
-#include "fastepd_stock.h"
-namespace fastepd = fastepd_stock;
-#endif
+#include <FastEPD.h>
+#include <stdint.h>
+
+extern FASTEPD bbep;
+
+namespace fastepd {
+  void init();
+  void update(bool bWait, bool bSkipClear, int iUpdateCount, uint32_t iTempProfile);
+} // namespace fastepd
+
+#endif // BOARD_X_CLASS
